+++ apps/mobile-wallet/src/features/modals/useIsTopModal.tsx
@@ -1,0 +1,11 @@
+import { usePotentialModalId } from '~/features/modals/ModalContext'
+import { selectIsTopModal } from '~/features/modals/modalSelectors'
+import { useAppSelector } from '~/hooks/redux'
+
+const useIsTopModal = () => {
+  const modalId = usePotentialModalId()
+
+  return useAppSelector((state) => !!modalId && selectIsTopModal(state, modalId))
+}
+
+export default useIsTopModal